+++ tests/testthat/test-deduped_map.R
@@ -1,42 +1,8 @@
-test_that("deduped_map(f) runs only on deduplicated values", {
-  get_ncalls <- function() as.integer(Sys.getenv("ncalls"))
-  increment_ncalls <- function() Sys.setenv(ncalls = get_ncalls() + 1L)
-
-  f <- function(i) increment_ncalls()
-
-  x <- list(
-    c(1, 2),
-    "A",
-    data.frame(x = "A"),
-    data.frame(y = 1)
+test_that("deduped_map warns for deprecation, but works", {
+  x <- list("ABC", "ABC")
+  expect_warning(deduped_map(x, tolower), "deprecated")
+  expect_equal(
+    suppressWarnings(deduped_map(x, tolower)),
+    deduped(lapply)(x, tolower)
   )
-  xr <- rep(x, 5)
-
-  withr::with_envvar(
-    c(ncalls = 0),
-    {
-      deduped_map(xr, f)
-      expect_equal(get_ncalls(), length(x))
-    }
-  )
-
-  withr::with_envvar(
-    c(ncalls = 0),
-    {
-      deduped_map(x, f)
-      expect_equal(get_ncalls(), length(x))
-    }
-  )
-})
-
-test_that("deduped_map(f) returns the data in the same order", {
-  x <- list(
-    c(1, 2),
-    "A",
-    data.frame(x = "A"),
-    data.frame(y = 1)
-  )
-  xr <- rep(x, 5)
-  pass_through <- function(i) i
-  expect_equal(deduped_map(xr, pass_through), xr)
 })